OPA fmt + lint
