+++ classification/labels/ccn.rego
@@ -2,6 +2,8 @@ package classifier_ccn
 
 import rego.v1
 
+# METADATA
+# entrypoint: true
 output[k] := v if {
 	some k in object.keys(input)
 	v := classify(k, input[k])
@@ -12,6 +14,6 @@ default classify(_, _) := false
 classify(_, val) if {
 	regex.match(
 		`\A(?:4[0-9]{12}(?:[0-9]{3})?|[25][1-7][0-9]{14}|6(?:011|5[0-9][0-9])[0-9]{12}|3[47][0-9]{13}|3(?:0[0-5]|[68][0-9])[0-9]{11}|(?:2131|1800|35\d{3})\d{11})\z`,
-		val
-    )
+		val,
+	)
 }
